verify that the zipfile doesnt try to escape up

+++ initializr.py
@@ -40,16 +40,23 @@ class InitializrProject(sublime_plugin.WindowCommand):
         if os.path.exists(path):
             self.window.show_input_panel('That path exists! Where to save your new project?',
                                          '~/newproject', self.create, None, None)
-        self.download(None)
+        self.download()
+        if not self.verify():
+            sublime.error_message('Zipfile contained a suspicious path. Aborting...')
+            return
+
         mkdir_p(path)
         self.zf.extractall(path)
-        for fn in glob.glob('%s/initializr/*' % path):
-            shutil.move(fn, '%s/' % path)
-        os.rmdir('%s/initializr' % path)
+        try:
+            for fn in glob.glob('%s/initializr/*' % path):
+                shutil.move(fn, '%s/' % path)
+            os.rmdir('%s/initializr' % path)
+        except:
+            pass
 
         self.finish(path)
 
-    def download(self, url):
+    def download(self):
         if not url:
             url = self.settings.get('zip_url')
         try:
@@ -57,6 +64,12 @@ class InitializrProject(sublime_plugin.WindowCommand):
         # For ST2 Compat
         except AttributeError:
             self.zf = ZipFile(BytesIO(open_url(url).read()))
+
+    def verify(self):
+        for name in self.zf.namelist():
+            if '../' in name:
+                return False
+        return True
 
     def finish(self, path):
         project_data = {
